Исправить замечания
- все id Long
- переопределены equals и hashcode
- имя роли уникально
- удалены @Repository
- удалены неиспользуемые методы в RoleService
- роли подтягиваются через join fetch
- убраны лишние пробелы
- в AdminController инжект через интерфейс
- применены PATCH DELETE маппинги

+++ src/main/java/ru/kata/spring/boot_security/demo/services/UserService.java
@@ -1,7 +1,4 @@
 package ru.kata.spring.boot_security.demo.services;
-
-
-
 import ru.kata.spring.boot_security.demo.models.User;
 
 import java.util.List;
@@ -10,13 +7,13 @@ import java.util.List;
 public interface UserService {
     List<User> getAllUsers();
 
-    User findUserById(int id);
+    User findUserById(Long id);
 
     User findUserByEmail(String email);
 
     void addUser(User user);
 
-    void updateUser(int id, User updatedUser);
+    void updateUser(Long id, User updatedUser);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 }
