+++ src/main/java/ru/kata/spring/boot_security/demo/services/UserServiceImp.java
@@ -25,47 +25,39 @@ public class UserServiceImp implements UserService {
     @Override
     @Transactional(readOnly = true)
     public List<User> getAllUsers() {
-
         return userRepository.findAll();
     }
 
     @Override
     @Transactional(readOnly = true)
-    public User findUserById(int id) {
-
+    public User findUserById(Long id) {
         return userRepository.findById(id).orElse(null);
     }
 
     @Override
     @Transactional(readOnly = true)
     public User findUserByEmail(String email) {
-
         return userRepository.findByEmail(email).orElse(null);
     }
 
     @Override
     @Transactional
     public void addUser(User user) {
-
         user.setPassword(passwordEncoder.encode(user.getPassword()));
-
         userRepository.save(user);
     }
 
     @Override
     @Transactional
-    public void updateUser(int id, User updatedUser) {
-
+    public void updateUser(Long id, User updatedUser) {
         updatedUser.setPassword(passwordEncoder.encode(updatedUser.getPassword()));
         updatedUser.setId(id);
-
         userRepository.save(updatedUser);
     }
 
     @Override
     @Transactional
-    public void deleteUser(int id) {
-
+    public void deleteUser(Long id) {
         userRepository.deleteById(id);
     }
 }
